+++ src/pages/deriv-go/_footer-banner.tsx
@@ -6,7 +6,7 @@ import GrayAngle30 from '../../images/common/gray-angle.png'
 import AppleIcon from '../../images/svg/apple_icon.svg'
 import AppGalleryIcon from '../../images/svg/appGallery_icon.svg'
 import AndroidIcon from '../../images/svg/android_icon.svg'
-import GetAppMobileBG from '../../images/common/dtrader/getAppMobileBG.jpg'
+import GetAppMobileBG from '../../images/common/dtrader/get_app_mobile_bg.jpg'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import MultiWidthColumn from 'components/elements/multi-width-column'
 import device from 'themes/device'
