+++ src/pages/derivx/_get-derivx.tsx
@@ -5,7 +5,7 @@ import AppleIcon from '../../images/svg/apple_icon.svg'
 import AppGalleryIcon from '../../images/svg/appGallery_icon.svg'
 import AndroidIcon from '../../images/svg/android_icon.svg'
 import BrowserIcon from '../../images/svg/browser_icon.svg'
-import GetAppMobileBG from '../../images/common/getAppMobileBG.png'
+import GetAppMobileBG from '../../images/common/get_app_mobile_bg.png'
 import { DerivXLogo } from 'images/svg/deriv-x'
 import DerivXQR from 'images/common/deriv-x/derivxQR.png'
 import CommonHeaderSection from 'components/elements/common-header-section'
