+++ src/pages/dmt5/_download-app.tsx
@@ -11,7 +11,7 @@ import BrowserIcon from '../../images/svg/browser_icon.svg'
 import LinuxIcon from '../../images/svg/linux_icon.svg'
 import MacIcon from '../../images/svg/mac_icon.svg'
 import WinIcon from '../../images/svg/windows_icon.svg'
-import GetAppMobileBG from '../../images/common/getAppMobileBG.png'
+import GetAppMobileBG from '../../images/common/get_app_mobile_bg.png'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import MultiWidthColumn from 'components/elements/multi-width-column'
 import device from 'themes/device'
