+++ src/pages/dtrader/_get-app-section.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import derivTraderLogo from '../../images/common/dtrader/dtrader.svg'
 import GrayAngle30 from '../../images/common/gray-angle.png'
-import GetAppMobileBG from '../../images/common/dtrader/getAppMobileBG.jpg'
+import GetAppMobileBG from '../../images/common/dtrader/get_app_mobile_bg.jpg'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import MultiWidthColumn from 'components/elements/multi-width-column'
 import useHandleSignup from 'components/hooks/use-handle-signup'
